Create Formatting Facet
